requirements.txt
Django==1.9.8
ffmpeg-python>=0.2.0
